CancelOrdersByClientIds: error on dex failure
This instruction currently returns Ok even if an order cancellation
fails. In most cases this is the desired behavior, since the most common
error is that the order no longer exists. However, for other errors, we
should return the error instead of Ok since the cancellation removes the
order from the order book, and that partial state update is no longer
valid.